2026-4-28 FloatField 更改符合 溫度欄位

+++ myapp/models.py
@@ -8,7 +8,7 @@ class Temperature(models.Model):
     #timestamp datatime
     myid = models.AutoField(primary_key=True)
     sensor_id = models.IntegerField(blank=False, null=False)
-    temperature= models.IntegerField(blank=False, null=False)
+    temperature= models.FloatField(blank=False, null=False)
     humidity = models.FloatField(blank=False, null=False)
     timestamp = models.DateTimeField(auto_now=True)
     def __str__(self):
